requirements.txt
grpcio
grpcio-tools
tifffile==2022.10.10
requests==2.31.0
keras==3.2.1
tensorflow==2.16.1
imagecodecs
geopandas==1.0.1
shapely
gdown
dill==0.3.8
flask==3.0.3
pytest
pylance
pylint
black
flake8
folium==0.17.0
pillow==11.0.0
matplotlib==3.9.2
